refactor: extract "ArticleViewer" and mapping logic for favicons from /blog page
