+++ packages/apps/web/src/components/article-viewer/index.ts
@@ -1,0 +1,1 @@
+export * from '#/components/article-viewer/ArticleViewer';
